+++ docs/changelog.md
@@ -1,5 +1,9 @@
 # Changelog
 
+### x.y.z - Bugfix
+
+ - Fixed `IndexError` when `NaN` values are present in the dataframe. Fixes [#28](https://github.com/python-qds/qdscreen/issues/28) 
+
 ### 0.6.2 - Warning filter
 
  - Now filtering `UserWarning` in `fit_selector_model` even in the sklearn adapter. Fixes [#20](https://github.com/python-qds/qdscreen/issues/20)