+++ docs/changelog.md
@@ -1,6 +1,6 @@
 # Changelog
 
-### x.y.z - Bugfixes
+### 0.6.3 - Bugfixes
 
  - Fixed `ValueError` with recent versions of `SciPy`, due to usage of sparse arrays with object dtype. Fixes [#31](https://github.com/python-qds/qdscreen/issues/31)
  - Fixed `IndexError` when `NaN` values are present in the dataframe. Fixes [#28](https://github.com/python-qds/qdscreen/issues/28)